Move validateSegmentatation to tests/cpp/validator.h.
