+++ tests/cpp/utils.h
@@ -722,12 +722,6 @@ bool isSchedulerInUse(
 // Disable magic zero
 constexpr CompileParams matmul_cparams{DataType::Int32, 255, false};
 
-// Validate that the fusion is segmented with desired scheduler, currently only
-// supporting two segments
-void validateSegmentation(
-    FusionKernelRuntime* runtime,
-    const std::vector<ScheduleHeuristic>& expected_heuristics);
-
 // Utility to generate tensor with bias applied on the input tensor
 TensorView* biasEpilogue(TensorView* tensor, TensorView* bias);
